% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEGAHC.R
\name{CEG.AHC}
\alias{CEG.AHC}
\title{The AHC algorithm as written in Barclay 2013 
Author: Lorna Barclay}
\usage{
CEG.AHC(exampledata = exampledata, equivsize = 3)
}
\arguments{
\item{exampledata}{dataframe}
}
\description{
The AHC algorithm as written in Barclay 2013 
Author: Lorna Barclay
}
\examples{
CEG.AHC()

}
\keyword{ceg}
\keyword{model}
\keyword{selection,}
