% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_batch_monitor.R
\name{ceg.batch.monitor}
\alias{ceg.batch.monitor}
\title{A function to compute the batch monitor of a CEG}
\usage{
ceg.batch.monitor(df, struct, stage.key, stages, which.cut)
}
\arguments{
\item{df}{data in question}
}
\description{
A function to compute the batch monitor of a CEG
}
\examples{
 cegb.prior <- get.ref.prior(df, cegb.struct, cuts, cegb.stage.key, cegb.stages)
 bn.batch.monitor()
}
\keyword{batch}
\keyword{cut}
