% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_pa_child_monitor.R
\name{ceg.child.parent.monitor}
\alias{ceg.child.parent.monitor}
\title{A function to compute the node monitors of the BN}
\usage{
ceg.child.parent.monitor(df, target.stage, target.cut, condtnl.stage,
  stage.key, struct, stages, prior, n = 50, learn = FALSE)
}
\arguments{
\item{df}{data in question}

\item{target.stage}{child stage}

\item{target.cut}{which cut is the child stage in}

\item{stage.key}{ceg stage.key}

\item{struct}{ceg structure}

\item{stages}{list of stages}

\item{n}{number of iterations}

\item{learn}{binary to control learning}
}
\description{
A function to compute the node monitors of the BN
}
\examples{
chds.expect.pach <- ceg.child.parent.monitor(df=chds.df,
target.stage = "u1",
target.cut = 2, 
condtnl.stage = "u0",
struct = chds.struct,
stage.key = chds.sk.nocol, 
stages=chds.stages,
prior=chds.prior.expert,
n = 860,
learn = F)
}
\keyword{bn}
\keyword{monitor}
\keyword{node}
