% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_monitor.R
\name{component.monitor}
\alias{component.monitor}
\title{A function that returns the component monitors}
\usage{
component.monitor(data, prior)
}
\arguments{
\item{data}{from get.data fn}

\item{prior}{effective sample size over number of edges}
}
\description{
A function that returns the component monitors
}
\examples{
expct.cnts <- expected.counts(chds.prior, chds.which.stage)
expctBF <-component.monitor(expct.cnts,chds.prior)

 
}
\keyword{BayesFactor}
