% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_monitor.R
\name{expected.counts}
\alias{expected.counts}
\title{Returns the expected counts}
\usage{
expected.counts(prior, which.stage)
}
\arguments{
\item{prior}{effective sample size divided by number of outgoing edges}

\item{which.stage}{stage under consideration}

\item{which.cut}{cut containing the stage in question}
}
\description{
Returns the expected counts
}
\examples{
chds.which.stage <- c(1,2,2,3,3,4,4,4)
expct.cnts <- expected.counts(chds.prior, chds.which.stage)
}
\keyword{BayesFactor}
