% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getprior.R
\name{get.ref.prior}
\alias{get.ref.prior}
\title{A function to set the reference prior on a CEG}
\usage{
get.ref.prior(df, struct, stage.key)
}
\arguments{
\item{df}{data in question}

\item{struct}{ceg struct}

\item{stage.key}{which arrows come from which stages}
}
\description{
A function to set the reference prior on a CEG
}
\examples{
radical.prior <-get.ref.prior(df=radical.df,struct=radical.struct,cuts=radical.cuts,stage.key=radical.sk,stages=radical.stages)#check that we can get the prior
}
\keyword{prior}
