% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_monitor.R
\name{getdata}
\alias{getdata}
\title{A function to compute the component monitor of a CEG}
\usage{
getdata(df, stage.key)
}
\arguments{
\item{df}{data in question}

\item{stage.key}{ceg stage leu}

\item{struct}{ceg.struct}

\item{stages}{list of stages}

\item{which.cut}{which cut to examine}
}
\description{
A function to compute the component monitor of a CEG
}
\examples{
getdata(radical.df, radical.sk.nocol) -> radical.data
}
\keyword{leave}
\keyword{one}
\keyword{out}
\keyword{prequential}
