% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_ceg.R
\name{getTreeRenderable}
\alias{getTreeRenderable}
\title{A function to translate a stage.key to a renderable data.tree structure}
\usage{
getTreeRenderable(stage.key, df)
}
\arguments{
\item{stage.key}{which arrows come from which stages}

\item{df}{data in question}
}
\description{
A function to translate a stage.key to a renderable data.tree structure
}
\examples{
getTreeRenderable(stage.key, df)
}
\keyword{ceg,}
\keyword{interactive}
\keyword{viz,}
