% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_monitor.R
\name{part.monitor}
\alias{part.monitor}
\title{The partition monitor examines the probability p(U_t | x^{t-1}), 
i.e. observing the partitoning into stagings for the elements in a cut}
\usage{
part.monitor(rho, epsilon, df_cut, which.cut, stage.key, n.monitor,
  k = 0.8, tau = 5)
}
\arguments{
\item{rho}{double of the probability of situations composing stage changing}

\item{epsilon}{integer of maximum allowable distances on the hasse diagram}

\item{df_cut}{data frame}

\item{which.cut}{integer of the cut that we want to check the staging of}

\item{stage.key}{list of data frames generated by tostagekey()}

\item{n.monitor}{integer of number of records to consider from the data frame}

\item{k}{is some parameter between 0 and 1 for downweighting past data}

\item{tau}{double of the rate of forgetting}
}
\description{
The partition monitor examines the probability p(U_t | x^{t-1}), 
i.e. observing the partitoning into stagings for the elements in a cut
}
\examples{

}
\keyword{forecast,}
\keyword{monitor}
\keyword{one}
\keyword{stage,}
\keyword{step}
