% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.message.R
\name{pass.message}
\alias{pass.message}
\title{Message passing algorithm as outlined by Collazo 2017
takes a well ordered CEG and C-copmatible information I
outputs: an uncolored CEG with pi_hat}
\usage{
pass.message(df, stage.key, evidence, prior, stages, struct)
}
\arguments{
\item{df}{data frame}

\item{stage.key}{}

\item{evidence}{chunk of data frame representing the root to sink paths}

\item{prior}{Dirichlet prior with effective sample size as the largest number of outgoing edges}

\item{stages}{vector of stage names}

\item{struct}{list of observed counts in each stage}
}
\description{
Message passing algorithm as outlined by Collazo 2017
takes a well ordered CEG and C-copmatible information I
outputs: an uncolored CEG with pi_hat
}
\examples{


}
\keyword{ceg}
\keyword{message}
\keyword{passing,}
