% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_ceg.R
\name{renderCEG}
\alias{renderCEG}
\title{A function to automatically visualize the CEG}
\usage{
renderCEG(stage.key, df)
}
\arguments{
\item{stage.key}{which arrows come from which stages}
}
\description{
@param df data in question
}
\keyword{ceg,}
\keyword{viz}
