% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveoneout.R
\name{sit.resids.plot}
\alias{sit.resids.plot}
\title{returns the situation residual plot}
\usage{
sit.resids.plot(u, level)
}
\arguments{
\item{u}{stage}

\item{level}{edge of interest}
}
\description{
returns the situation residual plot
}
\examples{
sit.resids.plot(8,2,chds.loo.counts, chds.data,chds.prior)

}
\keyword{BayesFactor}
