% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveoneout.R
\name{sit.resids.plot2}
\alias{sit.resids.plot2}
\title{possibly a duplicate function i do not need}
\usage{
sit.resids.plot2(u, level, loo.counts, data, prior)
}
\arguments{
\item{u}{stage}

\item{level}{level of interest}

\item{data}{dataset}

\item{prior}{effective sample size divided by number of outgoing edges}
}
\description{
possibly a duplicate function i do not need
}
\examples{
sit.resids.plot2(8,2,chds.loo.counts, chds.data,chds.prior)

}
\keyword{BayesFactor}
