% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveoneout.R
\name{surprise.score}
\alias{surprise.score}
\title{computes the surprise score}
\usage{
surprise.score(u, loo.counts, data, prior)
}
\arguments{
\item{u}{#stage}

\item{loo.counts}{#list containing counts with each one left out}

\item{data}{#dataset}

\item{prior}{#ceg prior}
}
\description{
computes the surprise score
}
\examples{
surprise.score(8)
map(c(4,6:8),surprise.score)
}
\keyword{position}
\keyword{stage}
