% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveoneout.R
\name{surprise.score2}
\alias{surprise.score2}
\title{an alternative way to get the surprise score??}
\usage{
surprise.score2(u)
}
\arguments{
\item{u}{stage in quesiton}
}
\description{
an alternative way to get the surprise score??
}
\examples{
surprise.score2(8)
map(c(4,6:8),surprise.score2)
}
