% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostagekey.R
\name{to.struct}
\alias{to.struct}
\title{A function that can take the output from CEG.AHC() and put it into a list of data frames}
\usage{
to.struct(df, stage.key, sst)
}
\arguments{
\item{df}{dataframe}

\item{stage.key}{output of tostagekey()}

\item{sst}{output of CEG.AHC()}
}
\description{
A function that can take the output from CEG.AHC() and put it into a list of data frames
}
\examples{
 to.struct(chds.df,chds.stage.key,chds.sst) 

}
\keyword{ceg,}
\keyword{structure}
