% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostagekey.R
\name{tostagekey}
\alias{tostagekey}
\title{A function that can take the output from CEG.AHC() and put it into a list of data frames}
\usage{
tostagekey(df, sst)
}
\arguments{
\item{df}{dataframe}

\item{sst}{output of CEG.AHC()}
}
\description{
A function that can take the output from CEG.AHC() and put it into a list of data frames
}
\examples{
tostagekey(chds.df, chds.sst)-> chds.stage.key
}
\keyword{ceg,}
\keyword{structure}
