% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_monitor.R
\name{getdata}
\alias{getdata}
\title{A function to compute the component monitor of a CEG}
\usage{
getdata(df, stage.key)
}
\arguments{
\item{df}{data in question}
}
\description{
A function to compute the component monitor of a CEG
}
\keyword{batch}
\keyword{cut}
